requirements.txt
pytorch==1.7.1+cu110
torchvision==0.8.2+cu110
numpy==1.19.5
matplotlib
scikit-image
pandas
sklearn
progress